Fixes #1967 for windows-xul
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1855 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 6dc2f5acdd2c3654654cef956a1b115ad27e6d96

+++ tv/platform/windows-xul/xul/chrome/content/preffuncs.js
@@ -105,7 +105,7 @@ function hasMinSpaceChange() {
 
 function minSpaceChange() {
   var textbox = document.getElementById("minspace");
-  var value = parseInt(textbox.value);
+  var value = parseFloat(textbox.value);
   if ((value == 0) || (isNaN(value))) value = 1;
   textbox.value=value;
   pybridge.setPreserveDiskSpaceAmount(value);
